+++ core/src/de/tum/cit/ase/maze/Trap.java
@@ -13,7 +13,7 @@ public class Trap extends GameObject{
      * @param x The initial x-coordinate of the Trap.
      * @param y The initial y-coordinate of the Trap.
      */
-    public Trap(int x, int y) {
+    public Trap(float x, float y) {
         super(x, y);
         this.isReady = true; // Traps are ready by default
     }
